+++ package.json
@@ -2,8 +2,8 @@
   "name": "@fastify/one-line-logger",
   "version": "1.0.0",
   "description": "A simple formatter for fastify's logs",
-  "main": "src/index.js",
-  "types": "src/index.d.ts",
+  "main": "index.js",
+  "types": "types/index.d.ts",
   "directories": {
     "doc": "docs",
     "example": "examples",
